Fix: 댓글기능 오류 수정 및 컴포넌트 세분화
