+++ src/components/seungchanKim/LoginMain.js
@@ -3,6 +3,108 @@ import { Link } from 'react-router-dom';
 import '../../pages/seungchanKim/Login/login.scss';
 
 class LoginMain extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      id: '',
+      pw: '',
+      buttonChange: 'buttonEvent',
+      idFontSize: 'callNumber',
+      pwFontSize: 'placeHoldPw',
+      pwLook: 'pwLook',
+      type: 'password',
+      typeInfo: '비밀번호 표시',
+    };
+  }
+
+  handleLogin = () => {
+    fetch('http://10.58.0.98:8000/users/signin', {
+      method: 'POST',
+      body: JSON.stringify({
+        email: this.state.id,
+        password: this.state.pw,
+      }),
+    })
+      .then(response => response.json())
+      .then(response => {
+        if (response.message === 'SUCCESS') {
+          alert(`Welcome ${this.state.id}`);
+          this.props.history.push('/main-chan');
+        } else {
+          alert('Login failed');
+        }
+      })
+      .catch(err => alert(err));
+  };
+
+  pwClick = () => {
+    if (this.state.typeInfo === '비밀번호 표시') {
+      this.setState({
+        typeInfo: '숨기기',
+        type: 'text',
+      });
+    } else if (this.state.typeInfo === '숨기기') {
+      this.setState({
+        typeInfo: '비밀번호 표시',
+        type: 'password',
+      });
+    }
+  };
+
+  fontSize = () => {
+    if (this.state.id.length > 0) {
+      this.setState({
+        idFontSize: 'callNumberChange',
+      });
+    } else {
+      this.setState({
+        idFontSize: 'callNumber',
+      });
+    }
+  };
+
+  pwfontSize = () => {
+    if (this.state.pw.length > 0) {
+      this.setState({
+        pwFontSize: 'placeHoldPwChange',
+        pwLook: 'pwLookChange',
+      });
+    } else {
+      this.setState({
+        pwFontSize: 'placeHoldPw',
+        pwLook: 'pwLook',
+      });
+    }
+  };
+
+  handleIdInput = e => {
+    this.setState({
+      id: e.target.value,
+    });
+  };
+
+  handlePwInput = e => {
+    this.setState({
+      pw: e.target.value,
+    });
+  };
+
+  loginKey = () => {
+    if (
+      this.state.id.length > 5 &&
+      this.state.id.includes('@') &&
+      this.state.pw.length > 5
+    ) {
+      this.setState({
+        buttonChange: 'buttonEventChange',
+      });
+    } else {
+      this.setState({
+        buttonChange: 'buttonEvent',
+      });
+    }
+  };
+
   render() {
     return (
       <div className="login_main">
@@ -14,11 +116,11 @@ class LoginMain extends React.Component {
                 <input
                   type="text"
                   id="cs_id"
-                  value={this.props.idValue}
-                  onChange={this.props.handleIdInput}
-                  onKeyUp={this.props.fontSize}
+                  value={this.state.id}
+                  onChange={this.handleIdInput}
+                  onKeyUp={this.fontSize}
                 />
-                <span className={this.props.idFontSize}>
+                <span className={this.state.idFontSize}>
                   전화번호, 사용자 이름 또는 이메일
                 </span>
               </label>
@@ -26,32 +128,33 @@ class LoginMain extends React.Component {
             <div
               id="customer_pw"
               className="login_box_layout"
-              onKeyUp={this.props.loginKey}
+              onKeyUp={this.loginKey}
             >
               <label for="cs_pw">
                 <input
-                  type={this.props.type}
+                  type={this.state.type}
                   id="cs_pw"
-                  value={this.props.pwValue}
-                  onChange={this.props.handlePwInput}
-                  onKeyUp={this.props.pwPlaceHold}
+                  value={this.state.pw}
+                  onChange={this.handlePwInput}
+                  onKeyUp={this.pwfontSize}
                 />
-                <span className={this.props.pwFontSize}>비밀번호</span>
+                <span className={this.state.pwFontSize}>비밀번호</span>
               </label>
-              <div className={this.props.pwLook} onClick={this.props.pwClick}>
-                {this.props.typeInfo}
+              <div className={this.state.pwLook} onClick={this.pwClick}>
+                {this.state.typeInfo}
               </div>
             </div>
             <div
               id="login_button"
               className="login_box_layout"
-              onKeyUp={this.props.loginKey}
+              onKeyUp={this.loginKey}
             >
-              <Link to="/main-chan">
+              <Link>
                 <input
                   type="button"
                   value="로그인"
-                  className={this.props.buttonChange}
+                  className={this.state.buttonChange}
+                  onClick={this.handleLogin}
                 />
               </Link>
             </div>
